change vsupply to be a uint16 to enable reporting of voltages higher than 25.5V

+++ sw/airborne/firmwares/fixedwing/main_ap.c
@@ -125,7 +125,7 @@ bool_t launch = FALSE;
 /** Supply voltage in deciVolt.
  * This the ap copy of the measurement from fbw
  */
-uint8_t vsupply;
+uint16_t vsupply;
 
 /** Supply current in milliAmpere.
  * This the ap copy of the measurement from fbw
